+++ Content.Server/Backmen/Surgery/Wounds/Systems/ServerWoundSystem.cs
@@ -121,90 +121,7 @@ public sealed class ServerWoundSystem : WoundSystem
         return actualHeal > 0;
     }
 
-    private void UpdateWoundableIntegrity(EntityUid uid, WoundableComponent? component = null)
-    {
-        if (!_woundableQuery.Resolve(uid, ref component, false) || component.Wounds == null)
-            return;
 
-        // Ignore scars for woundable integrity.. Unless you want to confuse people with minor woundable state
-        var damage =
-            component.Wounds.ContainedEntities.Select(_woundQuery.Comp)
-                .Where(wound => !wound.IsScar)
-                .Aggregate(FixedPoint2.Zero, (current, wound) => current + wound.WoundIntegrityDamage);
-
-        var newIntegrity = FixedPoint2.Clamp(component.IntegrityCap - damage, 0, component.IntegrityCap);
-        if (newIntegrity == component.WoundableIntegrity)
-            return;
-
-        var ev = new WoundableIntegrityChangedEvent(component.WoundableIntegrity, newIntegrity);
-        RaiseLocalEvent(uid, ref ev);
-
-        var bodySeverity = FixedPoint2.Zero;
-        var bodyPart = Comp<BodyPartComponent>(uid);
-
-        if (bodyPart.Body.HasValue)
-        {
-            var rootPart = Comp<BodyComponent>(bodyPart.Body.Value).RootContainer.ContainedEntity;
-            if (rootPart.HasValue)
-            {
-                bodySeverity =
-                    GetAllWoundableChildren(rootPart.Value)
-                        .Aggregate(bodySeverity, (current, woundable) => current + GetWoundableIntegrityDamage(woundable, woundable));
-            }
-
-            var ev1 = new WoundableIntegrityChangedOnBodyEvent(
-                (uid, component),
-                bodySeverity - (component.WoundableIntegrity - newIntegrity),
-                bodySeverity);
-            RaiseLocalEvent(bodyPart.Body.Value, ref ev1);
-        }
-
-        component.WoundableIntegrity = newIntegrity;
-        Dirty(uid, component);
-    }
-
-    private void CheckWoundableSeverityThresholds(EntityUid woundable, WoundableComponent? component = null)
-    {
-        if (!_woundableQuery.Resolve(woundable, ref component, false))
-            return;
-
-        var nearestSeverity = component.WoundableSeverity;
-        foreach (var (severity, value) in component.Thresholds.OrderByDescending(kv => kv.Value))
-        {
-            if (component.WoundableIntegrity >= component.IntegrityCap)
-            {
-                nearestSeverity = WoundableSeverity.Healthy;
-                break;
-            }
-
-            if (component.WoundableIntegrity < value)
-                continue;
-
-            nearestSeverity = severity;
-            break;
-        }
-
-        if (nearestSeverity != component.WoundableSeverity)
-        {
-            var ev = new WoundableSeverityChangedEvent(component.WoundableSeverity, nearestSeverity);
-            RaiseLocalEvent(woundable, ref ev);
-        }
-        component.WoundableSeverity = nearestSeverity;
-
-        Dirty(woundable, component);
-
-        var bodyPart = Comp<BodyPartComponent>(woundable);
-        if (bodyPart.Body == null)
-            return;
-
-        if (!TryComp<TargetingComponent>(bodyPart.Body.Value, out var targeting))
-            return;
-
-        targeting.BodyStatus = GetWoundableStatesOnBodyPainFeels(bodyPart.Body.Value);
-        Dirty(bodyPart.Body.Value, targeting);
-
-        RaiseNetworkEvent(new TargetIntegrityChangeEvent(GetNetEntity(bodyPart.Body.Value)), bodyPart.Body.Value);
-    }
 
     public override void Update(float frameTime)
     {
